requirements.txt
asgiref==3.8.1
attrs==23.2.0
black==24.4.2
cattrs==23.2.3
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.8
dj-database-url==2.1.0
Django==5.0.6
django-dotenv==1.4.2
django-extensions==3.2.3
django-phonenumber-field==7.3.0
django-schema-graph==3.1.0
exceptiongroup==1.2.1
filelock==3.14.0
identify==2.5.36
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==24.0
pathspec==0.12.1
phonenumbers==8.13.36
pillow==10.3.0
platformdirs==4.2.2
pre-commit==3.7.1
psycopg2==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
tzdata==2024.1
virtualenv==20.26.2
